# backend/internal/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EpicPanel API",
    "version": "1.0.0",
    "description": "Linux-first hosting control plane API. Session cookie or Bearer session token for users; Bearer epk_ tokens for automation (scoped)."
  },
  "servers": [{"url": "/"}],
  "components": {
    "securitySchemes": {
      "bearer": {"type": "http", "scheme": "bearer"}
    },
...
  },
  "paths": {
...
    "/v1/organizations": {
      "get": {"summary": "List organizations of current user", "security": [{"bearer": []}], "responses": {"200": {"description": "ok"}}},
      "post": {"summary": "Create organization (creator becomes owner)", "security": [{"bearer": []}], "responses": {"201": {"description": "created"}, "409": {"description": "slug taken"}}}
    },
    "/v1/organizations/{org_id}": {
      "get": {"summary": "Organization details (billing+)", "security": [{"bearer": []}], "responses": {"200": {"description": "ok"}, "404": {"description": "not a member"}}},
      "patch": {"summary": "Rename organization (admin+)", "security": [{"bearer": []}], "responses": {"200": {"description": "ok"}}}
    },
    "/v1/organizations/{org_id}/members": {
      "get": {"summary": "List members (billing+)", "security": [{"bearer": []}], "responses": {"200": {"description": "ok"}}},
      "post": {"summary": "Add member by email (admin+)", "security": [{"bearer": []}], "responses": {"201": {"description": "added"}, "404": {"description": "no such user"}}}
    },
    "/v1/organizations/{org_id}/members/{user_id}": {
      "patch": {"summary": "Change member role (admin+)", "security": [{"bearer": []}], "responses": {"204": {"description": "done"}}},
      "delete": {"summary": "Remove member (admin+, last owner protected)", "security": [{"bearer": []}], "responses": {"204": {"description": "done"}, "409": {"description": "last owner"}}}
    },
...
    "/v1/organizations/{org_id}/servers/{server_id}/runtimes": {
      "get": {"summary": "List runtimes (billing+)", "security": [{"bearer": []}], "responses": {"200": {"description": "ok"}}},
      "post": {"summary": "Install runtime {type, version} (admin+) -> 202", "security": [{"bearer": []}], "responses": {"202": {"description": "installing"}}}
    },
    "/v1/organizations/{org_id}/servers/{server_id}/runtimes/{runtime_id}": {"delete": {"summary": "Remove runtime (admin+; 409 while in use)", "security": [{"bearer": []}], "responses": {"202": {"description": "removing"}}}},
...
    "/v1/organizations/{org_id}/servers/{server_id}/detect-software": {"post": {"summary": "Inventory installed managed software and adopt it (admin+) -> 202", "security": [{"bearer": []}], "responses": {"202": {"description": "detecting"}}}},
    "/v1/organizations/{org_id}/servers/{server_id}/runtimes/{runtime_id}/extensions": {"get": {"summary": "List managed PHP extensions (billing+)", "security": [{"bearer": []}], "responses": {"200": {"description": "ok"}}}, "post": {"summary": "Install/remove a PHP extension {name, action} (developer+) -> 202", "security": [{"bearer": []}], "responses": {"202": {"description": "queued"}}}},
...
  }
}